BAARC-119: Move message queue client
