+++ internal/metamorph/async/mq_client.go
@@ -1,9 +1,7 @@
-package nats_mq
+package async
 
 import (
-	"fmt"
 	"log/slog"
-	"time"
 
 	"github.com/bitcoin-sv/arc/internal/metamorph"
 	"github.com/bitcoin-sv/arc/pkg/blocktx/blocktx_api"
@@ -20,40 +18,21 @@ const (
 	connectionTries = 5
 )
 
-type MQClient struct {
-	logger       *slog.Logger
-	nc           *nats.Conn
-	subscription *nats.Subscription
-	minedTxsChan chan *blocktx_api.TransactionBlocks
+type NatsClient interface {
+	QueueSubscribe(subj, queue string, cb nats.MsgHandler) (*nats.Subscription, error)
+	Close()
+	Publish(subj string, data []byte) error
+	Drain() error
 }
 
-func NewNatsMQClient(minedTxsChan chan *blocktx_api.TransactionBlocks, logger *slog.Logger, natsURL string) (metamorph.MessageQueueClient, error) {
+type MQClient struct {
+	nc           NatsClient
+	logger       *slog.Logger
+	minedTxsChan chan *blocktx_api.TransactionBlocks
+	subscription *nats.Subscription
+}
 
-	var nc *nats.Conn
-	var err error
-
-	nc, err = nats.Connect(natsURL)
-	if err == nil {
-		return &MQClient{nc: nc, logger: logger, minedTxsChan: minedTxsChan}, nil
-	}
-
-	// Try to reconnect in intervals
-	i := 0
-	for range time.NewTicker(2 * time.Second).C {
-		nc, err = nats.Connect(natsURL)
-		if err != nil && i >= connectionTries {
-			return nil, fmt.Errorf("failed to connect to NATS server: %v", err)
-		}
-
-		if err == nil {
-			break
-		}
-
-		logger.Info("Waiting before connecting to NATS", slog.String("url", natsURL))
-		i++
-	}
-
-	logger.Info("Connected to NATS at", slog.String("url", nc.ConnectedUrl()))
+func NewNatsMQClient(nc NatsClient, minedTxsChan chan *blocktx_api.TransactionBlocks, logger *slog.Logger) (metamorph.MessageQueueClient, error) {
 
 	return &MQClient{nc: nc, logger: logger, minedTxsChan: minedTxsChan}, nil
 }
@@ -104,6 +83,10 @@ func (c MQClient) Shutdown() error {
 	err := c.nc.Drain()
 	if err != nil {
 		return err
+	}
+
+	if c.minedTxsChan != nil {
+		close(c.minedTxsChan)
 	}
 
 	c.nc.Close()
